+++ src/org/ods/component/ScanWithSonarStage.groovy
@@ -80,6 +80,7 @@ class ScanWithSonarStage extends Stage {
         logger.startClocked("${sonarProjectKey}-sq-scan")
         scan(sonarProperties)
         logger.debugClocked("${sonarProjectKey}-sq-scan", (null as String))
+        retryComputeEngineStatusCheck()
 
         generateAndArchiveReports(
             sonarProjectKey,
@@ -202,6 +203,36 @@ class ScanWithSonarStage extends Stage {
         } catch (Exception ex) {
             steps.error 'Quality gate status could not be retrieved. ' +
                 "Status was: '${qualityGateJSON}'. Error was: ${ex}"
+        }
+    }
+
+    @TypeChecked(TypeCheckingMode.SKIP)
+    private String retryComputeEngineStatusCheck() {
+        def waitTime = 5 // seconds
+        def retries = 5
+        def taskProperties = sonarQube.readTask()
+
+        for (def i = 0; i < retries; i++) {
+            def computeEngineTaskResult
+            try {
+                computeEngineTaskResult = sonarQube.getComputeEngineTaskResult(taskProperties['ceTaskId'])
+            } catch (Exception ex) {
+                script.error 'Compute engine status could not be retrieved. ' +
+                "Status was: '${computeEngineTaskJSON}'. Error was: ${ex}"
+            }
+            if (computeEngineTaskResult == 'IN_PROGRESS' || computeEngineTaskResult == 'PENDING') {
+                logger.info "SonarQube background task has not finished yet."
+                script.sleep(waitTime)
+            } else if (computeEngineTaskResult == 'SUCCESS') {
+                logger.info "SonarQube background task has finished successfully."
+                break
+                } else if (computeEngineTaskResult == 'FAILED') {
+                logger.info "SonarQube background task has failed!"
+                steps.error 'SonarQube Scanner stage has ended with errors'
+                } else {
+                logger.info "Unknown status for the background task"
+                steps.error 'SonarQube Scanner stage has ended with errors'
+                }
         }
     }